+++ CHANGELOG.md
@@ -7,8 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+## [1.67.0](https://github.com/cloudogu/ces-build-lib/releases/tag/1.67.0) - 2023-09-04
+### Changed
+- Switch to hadolint Dockerfile linter; #111
 
-## [1.66.0](https://github.com/cloudogu/ces-build-lib/releases/tag/1.66.1) - 2023-09-04
+### Added
+- Add Dockerfile class with new linting functions; #111
+  - `lint()` lints with default hadolint config parameters. Only fails on errors
+  - `lintWithConfig()` lets you specify the hadolint configuration
+
+### Deprecated
+- `lintDockerfile()` function should be replaced by the Dockerfile `lint()` function
+
+## [1.66.1](https://github.com/cloudogu/ces-build-lib/releases/tag/1.66.1) - 2023-09-04
 ### Fixed
 - Split helm-repo-config in separate values #113